🚚 Move: src, app 폴더 경로 변경
- src/app/ 에 라우터와 페이지 관련된 파일들만 남기고 src/ 로 이동시켰습니다.
