+++ src/components/Todos/TodoListContainer.tsx
@@ -1,8 +1,8 @@
 'use client';
-import TodoLabel from '../../../../public/images/todo/todo.svg';
-import DoneLabel from '../../../../public/images/done/done.svg';
-import EmptyTodo_small from '../../../../public/images/empty/todo_small/Type=todo, Size=Small.svg';  
-import EmptyDone_small from '../../../../public/images/empty/done_small/Type=Done, Size=Small.svg';
+import TodoLabel from '../../../public/images/todo/todo.svg';
+import DoneLabel from '../../../public/images/done/done.svg';
+import EmptyTodo_small from '../../../public/images/empty/todo_small/Type=todo, Size=Small.svg';  
+import EmptyDone_small from '../../../public/images/empty/done_small/Type=Done, Size=Small.svg';
 import useTodoStore, { TodoProps } from '@/store/todoStore';
 import { useEffect, useState } from 'react';
 import TodoList from './TodoList';
